remove old ds18b20 sub modules

+++ main/contacts.h
@@ -1,7 +1,7 @@
 #ifndef __contacts_h
 #define __contacts_h
 
-typedef void (*contact_callback_t)(int value, void *userdata);
+typedef void (*contact_callback_t)(int8_t value, void *userdata);
 
 struct contact_s {
    int8_t last_state;
